style: Split into multiple spec files
