+++ packages/runtime-core/__tests__/renderList.spec.ts
@@ -1,28 +1,4 @@
-import { toHandlers, renderList } from '@vue/runtime-core'
-import { mockWarn } from '@vue/runtime-test'
-
-describe('toHandlers', () => {
-  mockWarn()
-
-  it('should not accept non-objects', () => {
-    toHandlers((null as unknown) as any)
-    toHandlers((undefined as unknown) as any)
-
-    expect(
-      'v-on with no argument expects an object value.'
-    ).toHaveBeenWarnedTimes(2)
-  })
-
-  it('should properly change object keys', () => {
-    const input = () => {}
-    const change = () => {}
-
-    expect(toHandlers({ input, change })).toStrictEqual({
-      oninput: input,
-      onchange: change
-    })
-  })
-})
+import { renderList } from '../src/helpers/renderList'
 
 describe('renderList', () => {
   it('should render nodes in an array', () => {
